fix(agents): one existence-specific death probe across rm and reap
Code-review round on PR 1601: process_start_time's None conflates a
dead process with a failed lookup, so death needed the ESRCH-only
probe; rm's live gate now accepts the same verdict the reaper does, or
a merge cleanup clears on evidence the next fno agents rm refuses;
the merge cleanup reads the roster once per pass, not once per row.

+++ crates/fno-agents/src/daemon/tests/gc_receipts.rs
@@ -3891,11 +3891,12 @@ fn dead_pid_proves_the_stop() {
             .push(claude_worker_row("row-pid", "dddd4444"));
     })
     .unwrap();
-    // u32::MAX cannot name a live process on any supported platform, so the
+    // i32::MAX cannot name a live process on any supported platform (and it
+    // casts to a positive pid_t, so kill reads it as an existence probe), so the
     // probe must answer "gone".
     let agents = crate::claude_roster::ClaudeAgentsSnapshot::known(vec![
         crate::claude_roster::ClaudeAgentRow::new("dddd4444", Some("working"))
-            .with_pid(Some(u32::MAX)),
+            .with_pid(Some(i32::MAX as u32)),
     ]);
     let stops = std::cell::RefCell::new(0usize);
     let summary = evidence_sweep(
